refactor(将dom相关函数移动至单独的文件):
